Moved plotting methods to new class
